move types to global types
