+++ src/features/problems/queries/getProblemScore.ts
@@ -1,23 +1,7 @@
 import prisma from "@database/prisma/instance";
 import type { DebugSubmissions } from "@prisma/client";
 
-import type { Results, UserData } from "@/types";
-
-interface DetailedScoreInfoNotNull {
-  score: number;
-  diff: number | null;
-  result: Results;
-  drid: number;
-}
-
-interface DetailedScoreInfoNull {
-  score: 0;
-  diff: null;
-  result: null;
-  drid: null;
-}
-
-export type DetailedScoreInfo = DetailedScoreInfoNotNull | DetailedScoreInfoNull;
+import type { DetailedScoreInfo, Results, UserData } from "@/types";
 
 export type GetProblemScoreResult =
   | { ok: true; data: DetailedScoreInfo }
